Make testing configurable and fix some bugs found with RBReakout.

+++ RRegistry/include/RRegistry/Entries.hpp
@@ -24,7 +24,7 @@ enum class Type
   _COUNT
 };
 
-typedef uint32_t EntryID;
+typedef uint16_t EntryID;
 
 enum class Int8
 {
